+++ pyodide-build/pyodide_build/buildpkg.py
@@ -484,6 +484,34 @@ def replace_so_abi_tags(wheel_dir: Path) -> None:
         file.rename(file.with_name(file.name.replace(build_triplet, host_triplet)))
 
 
+def copy_sharedlibs(
+    wheel_file: Path, wheel_dir: Path, lib_dir: Path
+) -> dict[str, Path]:
+    from auditwheel_emscripten import copylib, resolve_sharedlib  # type: ignore[import]
+    from auditwheel_emscripten.wheel_utils import WHEEL_INFO_RE  # type: ignore[import]
+
+    match = WHEEL_INFO_RE.match(wheel_file.name)
+    if match is None:
+        raise RuntimeError(f"Failed to parse wheel file name: {wheel_file.name}")
+
+    dep_map: dict[str, Path] = resolve_sharedlib(
+        wheel_dir,
+        lib_dir,
+    )
+    lib_sdir: str = match.group("name") + ".libs"
+
+    if dep_map:
+        dep_map_new = copylib(wheel_dir, dep_map, lib_sdir)
+        print("Copied shared libraries:")
+        for lib, path in dep_map_new.items():
+            original_path = dep_map[lib]
+            print(f"  {original_path} -> {path}")
+
+        return dep_map_new
+
+    return {}
+
+
 def package_wheel(
     pkg_name: str,
     pkg_root: Path,
@@ -537,6 +565,11 @@ def package_wheel(
     # update so abi tags after build is complete but before running post script
     # to maximize sanity.
     replace_so_abi_tags(wheel_dir)
+
+    vendor_sharedlib = build_metadata.vendor_sharedlib
+    if vendor_sharedlib:
+        lib_dir = Path(common.get_make_flag("WASM_LIBRARY_DIR"))
+        copy_sharedlibs(wheel, wheel_dir, lib_dir)
 
     post = build_metadata.post
     if post: